Decompose gofakes3.go
